+++ frontend/lib/widgets/brand-promotion/brand_promotion_widgets.dart
@@ -1,0 +1,3 @@
+export 'filter_button.dart';
+export 'promotion_block.dart';
+export 'floating_search_button.dart';
